requirements.txt
django
djangorestframework
drf-yasg
python-dotenv
django-grappelli
dj-email-url



